requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
requests==2.31.0
psycopg2-binary==2.9.9
